+++ arduino/libs/sleep/include/sleep.hpp
@@ -12,4 +12,5 @@ void rfNodeSleepAndPollForWakeup();
 void powerSaveSleepMs(uint8_t delay_ms);
 void powerDownRadioAndSleep(uint16_t delay);
 void commandGetActiveTimeCounter(uint8_t* commandPayload, uint8_t* responsePayload);
+void setAdvertisementInterval(uint32_t interval);
 } // namespace